tools/telegram-bot/requirements.txt
python-telegram-bot[ext]==21.6
httpx==0.27.0
qrcode[pil]==7.4.2
python-dotenv==1.0.1
